Demo selecting inventory items for Noxygenic
To answer the questions from Noxygenic at https://forum.botengine.org/t/need-help-with-selecting-items-in-inventory/3076?u=viir

+ Use `show info` on the matching inventory items.
+ Count the number of inventory items matching the filter and display this number.

----

Since I know the mining bot uses drag&drop, I can search the bot code for this operation to see where it reads the inventory item, and find this part:
https://github.com/Viir/bots/blob/3af6103a93a8a91d691f3fc15db45abf7203c827/implement/applications/eve-online/eve-online-mining-bot/src/Bot.elm#L110

For the new use case, we need to adapt this to take not the first item but to filter for `Sisters Core`. We could use the filtering function in the game UI, but this would require an additional step of entering the text in the text box. An easier way is to filter the inventory items in our app code.