Adapt to observation from TheRealManiac
From https://forum.botengine.org/t/mining-bot-question-and-hints/3140:

The Mining Barge is very slow and is the hole time approach behind the enemy, i change it in that way, that the target lock range, increased to 25000 meters.
if the enemy not in that range, i change it to “EndDecissionPath Wait” in the else branch from “acquireLockedTargetForCombat” an wait that the NPC came to my range.