Simplest change to fix and enable running the bot
Adapt the code copied from the example to the new version of the framework, so that we can run the bot.

After creation of the stop ship example, the framework on the main branch was expanded to allow multiple actions per reading from the game client, so there is a small change to wrap the action into a list.